Conflicts:
	Source/eu.modelwriter.model/src/eu/modelwriter/marker/model/samples/MasterView.java

+++ Source/eu.modelwriter.writer.markers/src/eu/modelwriter/marker/action/DeleteMarkerAction.java
@@ -7,12 +7,17 @@ import java.util.ArrayList;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.emf.ecore.ENamedElement;
+import org.eclipse.emf.ecore.presentation.EcoreEditor;
 import org.eclipse.jface.action.IAction;
 import org.eclipse.jface.dialogs.MessageDialog;
+import org.eclipse.jface.text.ITextSelection;
 import org.eclipse.jface.text.TextSelection;
 import org.eclipse.jface.viewers.ISelection;
+import org.eclipse.jface.viewers.ITreeSelection;
 import org.eclipse.ui.IEditorActionDelegate;
 import org.eclipse.ui.IEditorPart;
+import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.part.MultiPageEditorPart;
 import org.eclipse.ui.texteditor.IDocumentProvider;
 import org.eclipse.ui.texteditor.ITextEditor;
@@ -21,6 +26,7 @@ import eu.modelwriter.marker.Activator;
 import eu.modelwriter.marker.internal.MarkElement;
 import eu.modelwriter.marker.internal.MarkerFactory;
 import eu.modelwriter.marker.internal.Serialization;
+import eu.modelwriter.marker.xml.XMLDOMHelper;
 
 public class DeleteMarkerAction implements IEditorActionDelegate {
 
@@ -31,34 +37,76 @@ public class DeleteMarkerAction implements IEditorActionDelegate {
   @Override
   public void run(IAction action) {
     try {
-      TextSelection selection = MarkerFactory.getTextSelection();
-      IFile file = (IFile) Activator.getEditor().getEditorInput().getAdapter(IFile.class);
+      ISelection selection =
+          PlatformUI.getWorkbench().getActiveWorkbenchWindow().getSelectionService().getSelection();
 
-      IMarker beDeleted = MarkerFactory.findMarkerByOffset(file, selection.getOffset());
+      IFile file = (IFile) PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage()
+          .getActiveEditor().getEditorInput().getAdapter(IFile.class);
 
-      if (beDeleted != null) {
-        String markerText = (String) beDeleted.getAttribute(IMarker.TEXT);
+      IEditorPart editor =
+          PlatformUI.getWorkbench().getActiveWorkbenchWindow().getActivePage().getActiveEditor();
 
-        updateTargets(beDeleted);
-        updateSources(beDeleted);
+      IMarker beDeleted = null;
 
-        beDeleted.delete();
+      if (selection instanceof ITextSelection) {
+        TextSelection textSelection = (TextSelection) selection;
 
-        MessageDialog dialog =
-            new MessageDialog(Activator.getShell(), "Mark will be deleted by this wizard.", null,
-                "\"" + markerText + "\" has been seleceted to be unmarked",
+        beDeleted = MarkerFactory.findMarkerByOffset(file, textSelection.getOffset());
+        if (beDeleted != null && beDeleted.exists()) {
+          String markerText = (String) beDeleted.getAttribute(IMarker.TEXT);
+          updateTargets(beDeleted);
+          updateSources(beDeleted);
+
+          beDeleted.delete();
+
+          MessageDialog dialog =
+              new MessageDialog(Activator.getShell(), "Mark will be deleted by this wizard.", null,
+                  "\"" + markerText + "\" has been seleceted to be unmarked",
+                  MessageDialog.INFORMATION, new String[] {"OK"}, 0);
+          dialog.open();
+        }
+      } else if (selection instanceof ITreeSelection) {
+        ITreeSelection treeSelection = (ITreeSelection) selection;
+        if (selection != null && editor instanceof EcoreEditor
+            && treeSelection.getFirstElement() instanceof ENamedElement
+            && ((ENamedElement) treeSelection.getFirstElement()).getName() != null
+            && !((ENamedElement) treeSelection.getFirstElement()).getName().isEmpty()) {
+
+          String selectedText = ((ENamedElement) treeSelection.getFirstElement()).getName();
+
+          String xpath = XMLDOMHelper.findNodeAndGetXPath(selectedText,
+              file.getLocation().toFile().getAbsolutePath());
+
+          beDeleted = MarkerFactory.findMarkerByXpath(file, xpath);
+          if (beDeleted != null && beDeleted.exists()) {
+            updateTargets(beDeleted);
+            updateSources(beDeleted);
+
+            beDeleted.delete();
+            MessageDialog dialog =
+                new MessageDialog(Activator.getShell(), "Mark will be deleted by this wizard", null,
+                    "\"" + ((ENamedElement) ((ITreeSelection) selection).getFirstElement())
+                        .getName() + "\" has been seleceted to be unmarked",
                 MessageDialog.INFORMATION, new String[] {"OK"}, 0);
-        dialog.open();
+            dialog.open();
+          }
+        }
+      }
 
-        MultiPageEditorPart mpepEditor;
-        ITextEditor iteEditor;
-        if (Activator.getEditor() instanceof MultiPageEditorPart) {
-          mpepEditor = (MultiPageEditorPart) Activator.getEditor();
+      EcoreEditor ecEditor;
+      MultiPageEditorPart mpepEditor;
+      ITextEditor iteEditor = null;
+      if (editor instanceof EcoreEditor) {
+        ecEditor = (EcoreEditor) editor;
+        ecEditor.getViewer().refresh();
+      } else {
+        if (editor instanceof ITextEditor) {
+          iteEditor = (ITextEditor) editor;
+        } else {
+          mpepEditor = (MultiPageEditorPart) editor;
           IEditorPart[] editors = mpepEditor.findEditors(mpepEditor.getEditorInput());
           iteEditor = (ITextEditor) editors[0];
-        } else
-          iteEditor = (ITextEditor) Activator.getEditor();
-
+        }
         IDocumentProvider idp = iteEditor.getDocumentProvider();
         idp.resetDocument(iteEditor.getEditorInput());
       }
@@ -85,10 +133,10 @@ public class DeleteMarkerAction implements IEditorActionDelegate {
     try {
       if (marker.getAttribute(MarkElement.getTargetAttributeName()) != null) {
         ArrayList<MarkElement> targetElements = Serialization.getInstance() // güncellenen
-                                                                            // marker
-                                                                            // ın
-                                                                            // targetları
-                                                                            // alındı.
+            // marker
+            // ın
+            // targetları
+            // alındı.
             .fromString((String) (marker).getAttribute(MarkElement.getTargetAttributeName()));
 
         for (MarkElement targetElement : targetElements) {
